+++ source/Xamarin.Auth.LinkSource/AssemblyInfo.cs
@@ -24,7 +24,7 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Xamarin Inc.")]
 [assembly: AssemblyProduct("Xamarin.Auth")]
-[assembly: AssemblyCopyright("2012-2015 Xamarin Inc.")]
+[assembly: AssemblyCopyright("2012-2016 Xamarin Inc.")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
