+++ source/Xamarin.Auth.LinkSource/Xamarin.Auth.LinkSource.csproj
@@ -82,7 +82,9 @@
     <Compile Include="WebUtilities.cs" />
     <Compile Include="v.2\OAuth2Authenticator.HttpClient.cs" />
     <Compile Include="Authenticator.Portable.cs" />
-    <Compile Include="AccountStore.Async.cs" />
+    <Compile Include="AccountStore.Async.cs">
+      <DependentUpon>AccountStore.cs</DependentUpon>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
