+++ source/Xamarin.Auth.XamarinAndroid/Xamarin.Auth.XamarinAndroid.csproj
@@ -130,7 +130,9 @@
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="WebAuthenticator.XamarinAndroid.cs" />
     <Compile Include="AccountStore.XamarinAndroid.cs" />
-    <Compile Include="PlatformSpecific\AndroidAccountStore.Async.cs" />
+    <Compile Include="PlatformSpecific\AndroidAccountStore.Async.cs">
+      <DependentUpon>AndroidAccountStore.cs</DependentUpon>
+    </Compile>
     <Compile Include="..\Xamarin.Auth.LinkSource\AccountStore.Async.cs">
       <Link>AccountStore.Async.cs</Link>
     </Compile>
