+++ source/XamarinForms/Xamarin.Auth.XamarinForms.WindowsPhone8/Xamarin.Auth.XamarinForms.WindowsPhone8.csproj
@@ -126,10 +126,6 @@
       <Project>{08470E0D-EB43-4E07-92F4-020DF019F628}</Project>
       <Name>Xamarin.Auth.WindowsPhone8</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Xamarin.Auth.Helpers\Xamarin.Auth.Helpers.csproj">
-      <Project>{2cf4d47e-e4b0-454b-9015-1e76ee30c36e}</Project>
-      <Name>Xamarin.Auth.Helpers</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Xamarin.Auth.XamarinForms\Xamarin.Auth.XamarinForms.csproj">
       <Project>{67F9D3A8-F71E-4428-913F-C37AE82CDB24}</Project>
       <Name>Xamarin.Auth.XamarinForms</Name>
