+++ source/XamarinForms/Xamarin.Auth.XamarinForms.WindowsPhone81/Xamarin.Auth.XamarinForms.WindowsPhone81.csproj
@@ -125,10 +125,6 @@
       <Project>{18B5054D-6035-40DB-B1AE-3E4FDE8DB6E9}</Project>
       <Name>Xamarin.Auth.WindowsPhone81</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Xamarin.Auth.Helpers\Xamarin.Auth.Helpers.csproj">
-      <Project>{2cf4d47e-e4b0-454b-9015-1e76ee30c36e}</Project>
-      <Name>Xamarin.Auth.Helpers</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Xamarin.Auth.XamarinForms\Xamarin.Auth.XamarinForms.csproj">
       <Project>{67F9D3A8-F71E-4428-913F-C37AE82CDB24}</Project>
       <Name>Xamarin.Auth.XamarinForms</Name>
