+++ source/XamarinForms/Xamarin.Auth.XamarinForms.WindowsStore81WinRT/Xamarin.Auth.XamarinForms.WindowsStore81WinRT.csproj
@@ -141,10 +141,6 @@
       <Project>{2E9585E1-1917-4B6C-93E1-3446782E59EB}</Project>
       <Name>Xamarin.Auth.WindowsStore81WinRT</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Xamarin.Auth.Helpers\Xamarin.Auth.Helpers.csproj">
-      <Project>{2cf4d47e-e4b0-454b-9015-1e76ee30c36e}</Project>
-      <Name>Xamarin.Auth.Helpers</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Xamarin.Auth.XamarinForms\Xamarin.Auth.XamarinForms.csproj">
       <Project>{67F9D3A8-F71E-4428-913F-C37AE82CDB24}</Project>
       <Name>Xamarin.Auth.XamarinForms</Name>
